made the embedding storage stage multiprocess
