+++ src/nv_ingest/schemas/ingest_job_schema.py
@@ -89,8 +89,6 @@ class IngestTaskExtractSchema(BaseModelNoExt):
 
 
 class IngestTaskStoreEmbedSchema(BaseModelNoExt):
-    embedding: bool = True
-    method: str
     extra_params: dict
 
 class IngestTaskStoreSchema(BaseModelNoExt):
